requirements.txt
apscheduler==3.6.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.1.1
flask-cors==3.0.10
geopy==2.2.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.0
psycopg2==2.7.6.1
python-dateutil==2.9.0
requests==2.21.0
SQLAlchemy==1.3.0
urllib3==1.24.2
Werkzeug==0.16.1
./rws-common
